% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{adivboxplot}
\alias{adivboxplot}
\title{Alpha diversity boxplot}
\source{
[graphs.R](../R/graphs.R)
}
\usage{
adivboxplot(datafile, outdir, mapfile, amp = NULL, sampdepth = NULL,
  colors = NULL, cats = NULL, filesuffix = NULL, ...)
}
\arguments{
\item{datafile}{full path to input OTU file}

\item{outdir}{full path to output directory}

\item{mapfile}{full path to map file}

\item{amp}{ampvis2 object. may be specified instead of mapfile and datafile}

\item{sampdepth}{sampling depth.  see details.}

\item{colors}{colors to use for plots}

\item{cats}{categories/columns in mapping file to use as groups.  If NULL (default), will use
all columns starting with TreatmentGroup to (but not including) Description}

\item{filesuffix}{(Optional) suffix for output filename}

\item{...}{other parameters to pass to \link{readindata}}
}
\value{
Save alpha diversity boxplots to outdir.
}
\description{
Plots plotly boxplot of shannon diversity and Chao species richness.  If sampling depth is NULL,
rarefies OTU table to the minimum readcount of any sample.  If this is low, then the plot will fail.
}
\details{
If \code{sampdepth} is NULL, the sampling depth is set to the size of the smallest
sample.
}
