% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{allgraphs}
\alias{allgraphs}
\title{Pipeline function}
\source{
[graphs.R](../R/graphs.R)
}
\usage{
allgraphs(datafile, outdir, mapfile, sampdepth = 10000, ...)
}
\arguments{
\item{datafile}{full path to input OTU file (biom or txt file see \link{readindata} for format)}

\item{outdir}{full path to output directory}

\item{mapfile}{full path to map file}

\item{sampdepth}{sampling depth.  default: 10000}

\item{...}{other parameters to pass to \link{readindata}}
}
\value{
graphs are saved to outdir.  See [user doc](../doc/user_doc.md).

This value is used to remove samples before for alpha diversity and PCoA plots.
Also, to rarefy OTU table for the alpha diversity and Bray-Curtis distance PCoA.
}
\description{
Make all 4 types of graphs
}
