% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{amp_rarecurvefix}
\alias{amp_rarecurvefix}
\title{Rarefaction curve}
\source{
[utilities.R](../R/utilities.R)
}
\usage{
amp_rarecurvefix(data, stepsize = 1000, color_by = NULL)
}
\arguments{
\item{data}{(required) Data list as loaded with amp_load.}

\item{stepsize}{Step size for the curves. Lower is prettier but takes more time to generate. (default: 1000)}

\item{color_by}{Color curves by a variable in the metadata.}
}
\value{
A ggplot2 object.
}
\description{
This function replaces the ampvis2 function amp_rarecurve to fix subsampling labeling bug
in vegan
}
