% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{filterlowabund}
\alias{filterlowabund}
\title{Filter low abundant taxa}
\source{
[utilities.R](../R/utilities.R)
}
\usage{
filterlowabund(amp, level = 0.01, persamp = 0, abs = FALSE, toptaxa = NULL)
}
\arguments{
\item{amp}{ampvis2 object}

\item{level}{level at which to filter}

\item{persamp}{percent of samples which must have taxa in common}

\item{abs}{is level an absolute count? if false, will use level as relative percent.}

\item{toptaxa}{number of seqvar to include sorted by max count across all samples;
if NULL all will be included.}
}
\value{
filtered ampvis2 object
}
\description{
Filter low abundant taxa
}
