% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGrid.R
\name{gridCode}
\alias{gridCode}
\title{Format plotly grid code}
\source{
[plotlyGrid.R](../R/plotlyGrid.R)
}
\usage{
gridCode(data)
}
\arguments{
\item{data}{data to populate plotly grid}
}
\value{
list of 2 values:
\describe{
  \item{html}{html for plotly export link}
  \item{javascript}{js function for exporting data}
}
}
\description{
Format data according to here: \url{https://plot.ly/export/}
}
