% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{log10scale}
\alias{log10scale}
\alias{log10p_trans}
\title{Log base 10 + 1 scale}
\usage{
log10p_trans()
}
\value{
\code{log10p} returns a scales tranformation object
}
\description{
Transformation which computes \code{log10(x+1)} scale
}
\details{
\code{log10p} is for use with ggplot2 \code{trans} argument in scale function.
}
