% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{logoutput}
\alias{logoutput}
\title{write log output}
\source{
[utilities.R](../R/utilities.R)
}
\usage{
logoutput(c, bline = 0, aline = 0, type = NULL)
}
\arguments{
\item{c}{String. Log message/command to print.}

\item{bline}{Number of blank lines to precede output.}

\item{aline}{Number of blank lines to follow output.}

\item{type}{String.  Must be one of "WARNING", or "ERROR" or NULL.}
}
\description{
Prints time along with log message.
}
