% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{morphheatmap}
\alias{morphheatmap}
\title{Morpheus heatmap}
\source{
[graphs.R](../R/graphs.R)
}
\usage{
morphheatmap(datafile, outdir, mapfile, amp = NULL, sampdepth = NULL,
  rarefy = FALSE, filter_level = NULL, taxlevel = c("seq"),
  colors = NULL, rowAnnotations = NULL, force = FALSE,
  filesuffix = NULL, ...)
}
\arguments{
\item{datafile}{full path to input OTU file (biom or see \link{readindata})}

\item{outdir}{full path to output directory}

\item{mapfile}{full path to mapping file}

\item{amp}{(Optional) ampvis2 object. may be specified instead of mapfile and datafile}

\item{sampdepth}{sampling depth}

\item{rarefy}{Logical. Rarefy the OTU table if sampdepth is specified.}

\item{filter_level}{minimum abundance to show in the heatmap}

\item{taxlevel}{vector of taxonomic levels to graph.  must be subset of
c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species", "seq").  See Details.}

\item{colors}{(Optional) color vector - length equal to number of TreatmentGroups in mapfile}

\item{rowAnnotations}{(Optional) Row annotations to be used in addition to taxonomy.}

\item{force}{Force "seq" level heatmap to be made even if number of seqs is greater than 2000. '
See Details.}

\item{filesuffix}{(Optional) suffix for output filename}

\item{...}{parameters to pass to  \code{\link{readindata}}}
}
\value{
Saves heatmaps to outdir.
}
\description{
Creates heatmaps using Morpheus R API \url{https://software.broadinstitute.org/morpheus/}.  The heatmaps are made
using relative abundances.
}
\details{
For the \code{taxlevel} parameter, each level is made into a separate heatmap.  "seq" makes
 the heatmap with no collapsing of taxonomic levels if there are fewer than 2000 ASVs/OTUs.  Otherwise,
 Species level is made instead.
}
\examples{
\dontrun{
 morphheatmap(datafile="OTU_table.txt", outdir="outputs/graphs", mapfile="mapfile.txt",
 sampdepth = 25000, taxlevel = c("Family", "seq"), tsvfile=TRUE)
}


}
