% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{pcoaplot}
\alias{pcoaplot}
\title{PCoA plots}
\source{
[graphs.R](../R/graphs.R)
}
\usage{
pcoaplot(datafile, outdir, mapfile, amp = NULL, sampdepth = NULL,
  distm = "binomial", filter_species = 0.1, rarefy = FALSE,
  colors = NULL, filesuffix = NULL, ...)
}
\arguments{
\item{datafile}{full path to input OTU file (biom or see \link{readindata})}

\item{outdir}{full path to output directory}

\item{mapfile}{full path to map file}

\item{amp}{ampvis2 object. may be specified instead of mapfile and datafile}

\item{sampdepth}{sampling depth}

\item{distm}{distance measure for PCoA.  any that are supported by
\href{https://madsalbertsen.github.io/ampvis2/reference/amp_ordinate.html}{amp_ordinate} except for unifrac, wunifrac, and none.}

\item{filter_species}{Remove low abundant OTU's across all samples below this threshold in percent.
Setting this to 0 may drastically increase computation time.}

\item{rarefy}{Logical. Rarefy the OTU table if sampdepth is specified.}

\item{colors}{(Optional) color vector - length equal to number of TreatmentGroups in mapfile}

\item{filesuffix}{(Optional) suffix for output filename}

\item{...}{parameters to pass to  \code{\link{readindata}}}
}
\value{
Saves pcoa plots to outdir.
}
\description{
PCoA plots
}
