% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGrid.R
\name{plotlyGrid}
\alias{plotlyGrid}
\alias{htmlGrid}
\title{Add Plotly data export to Plotly graph}
\source{
[plotlyGrid.R](../R/plotlyGrid.R)
}
\usage{
plotlyGrid(pplot, filename, data = NULL, title = NULL, outlib = "lib")

htmlGrid(ht, filename, data, jquery = FALSE, title = NULL,
  outlib = "lib", styletags = NULL)
}
\arguments{
\item{pplot}{plotly or ggplot object}

\item{filename}{output filename (fullpath)}

\item{data}{data frame to export to plotly grid (optional for plotlyGrid)}

\item{title}{title of html page}

\item{outlib}{(Optional) name of external lib directory for non-selfcontained html.
Useful for multiple graphs sharing the same lib.}

\item{ht}{html tagList}

\item{jquery}{should we load jquery}

\item{styletags}{html object with style tags for the tagList.}
}
\value{
html plot is saved to filename. external libraries are saved to outlib in same directory as filename.
Invisibly returns the plotly html widget.
}
\description{
All functions create an output html plot  with link which sends the data to a grid in the plotly chart
 studio.

 \code{plotlyGrid} takes in a ggplot or plotly object and creates an output html plotly plot.

\code{htmlGrid} takes in an html tag object.
}
\details{
If jquery is needed, we use jquery-1.11.3 from the rmarkdown library.  We also use
 shiny's bootstrap-3.3.7 css to style the text elements.
}
