% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print_ampvis2}
\alias{print_ampvis2}
\title{Print ampvis2 object summary}
\source{
[utilities.R](../R/utilities.R)
}
\usage{
print_ampvis2(data)
}
\arguments{
\item{data}{ampvis2 object}
}
\value{
Prints summary stats about ampvis2 object
}
\description{
This is a copy of the internal ampvis2 function print.ampvis2.  CRAN does not allow
':::' internal calling of function in package.
}
