% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{rarefactioncurve}
\alias{rarefactioncurve}
\title{Make rarefaction curve graph}
\source{
[graphs.R](../R/graphs.R)
}
\usage{
rarefactioncurve(datafile, outdir, mapfile, amp = NULL, colors = NULL,
  cat = "TreatmentGroup", stepsize = 1000, ...)
}
\arguments{
\item{datafile}{full path to input OTU file (biom or see \link{readindata})}

\item{outdir}{full path to output directory}

\item{mapfile}{full path mapping file}

\item{amp}{(Optional) ampvis2 object. may be specified instead of mapfile and datafile}

\item{colors}{(Optional) color vector - length equal to number of TreatmentGroups in mapfile}

\item{cat}{Category/column in mapping file by which to color the curves in the graph.
(default TreatmentGroup)}

\item{stepsize}{for rarefaction plotting.}

\item{...}{parameters to pass to \code{\link{readindata}}}
}
\value{
Saves rarefaction curve plot to output directory.
}
\description{
Make rarefaction curve graph
}
