% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_biom}
\alias{read_biom}
\title{biomformat read_biom}
\usage{
read_biom(biom_file)
}
\arguments{
\item{biom_file}{input biom file name}
}
\value{
biom object
}
\description{
This function replaces the biomformat function read_biom to deal with reading in
crappy hdf5 biom file.
}
