% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{readindata}
\alias{readindata}
\title{Read in data}
\source{
[graphs.R](../R/graphs.R)
}
\usage{
readindata(datafile, mapfile, tsvfile = FALSE, mincount = 10)
}
\arguments{
\item{datafile}{full path to input data file.  must be either biom file or tab delimited text file.
See details.}

\item{mapfile}{full path to mapfile.  must contain SampleID, TreatmentGroup, and Description columns}

\item{tsvfile}{Logical.  Is datafile a tab-delimited text file? See details.}

\item{mincount}{minimum number of reads}
}
\value{
ampvis2 object
}
\description{
Read in data
}
\details{
datafile may be either biom file or text file.  If text file, it should have ampvis2 OTU table
format \url{https://madsalbertsen.github.io/ampvis2/reference/amp_load.html#the-otu-table}. If the
number of reads is less than mincount, the function will give an error, as we cannot make graphs
with so few counts.
}
