% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGrid.R
\name{save_fillhtml}
\alias{save_fillhtml}
\title{Save an HTML object to a file}
\source{
[plotlyGrid.R](../R/plotlyGrid.R)
}
\usage{
save_fillhtml(html, file, background = "white", libdir = "lib", bodystyle = "")
}
\arguments{
\item{html}{HTML content to print}

\item{file}{File to write content to}

\item{background}{Background color for web page}

\item{libdir}{Directory to copy dependencies to}

\item{bodystyle}{html style string}
}
\value{
save html to file
}
\description{
Save an HTML object to a file
}
