% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{shortnames}
\alias{shortnames}
\title{shortnames for taxonomy}
\source{
[utilities.R](../R/utilities.R)
}
\usage{
shortnames(taxtable)
}
\arguments{
\item{taxtable}{taxonomy table object from ampvis2 object amp$tax}
}
\value{
data.frame taxonomy table object like ampvis2 amp$tax.  taxonomy names
are sanitized and formatted to be a bit nicer.
}
\description{
shortnames for taxonomy
}
