% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{subsetamp}
\alias{subsetamp}
\title{Subset and rarefy OTU table.}
\source{
[graphs.R](../R/graphs.R)
}
\usage{
subsetamp(amp, sampdepth = NULL, rarefy = FALSE, printsummary = T,
  outdir = NULL, ...)
}
\arguments{
\item{amp}{ampvis2 object}

\item{sampdepth}{sampling depth.  See details.}

\item{rarefy}{rarefy the OTU table in addition to subsetting}

\item{printsummary}{Logical. print ampvis2 summary of OTU table}

\item{outdir}{Output directory.  If not null, and samples are removed from amp, the sample names will be output
to outdir/samples_being_ignored.txt}

\item{...}{other parameters to pass to amp_subset_samples}
}
\value{
ampvis2 object
}
\description{
Subset and/or rarefy OTU table.
}
\details{
\code{sampdepth} will be used to filter out samples with fewer than this number of reads.  If
rarefy is TRUE, then it will also be used as the depth at which to subsample using vegan function
rrarefy.
}
