% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{trygraphwrapper}
\alias{trygraphwrapper}
\title{Wrapper for any graph function}
\source{
[graphs.R](../R/graphs.R)
}
\usage{
trygraphwrapper(datafile, outdir, mapfile, FUN, logfilename = "logfile.txt",
  info = TRUE, tsvfile = FALSE, ...)
}
\arguments{
\item{datafile}{full path to input OTU file (biom or txt, see \link{readindata} for format of txt file)}

\item{outdir}{output directory for graphs}

\item{mapfile}{full path to map file}

\item{FUN}{character string. name of function you would like to run. can be actual
function object if run from R}

\item{logfilename}{logfilename}

\item{info}{print sessionInfo to logfile}

\item{tsvfile}{Is datafile a tab-delimited text file? Default FALSE}

\item{...}{parameters needed to pass to FUN}
}
\value{
Returns 0 if FUN succeeds and stops on error.  In rpy2, it will throw
rpy2.rinterface.RRuntimeError.
}
\description{
This is a wrapper for any of the graph functions meant to be called using rpy2 in python.
}
\examples{

\dontrun{

# example with no optional arguments for running allgraphs
trygraphwrapper("/path/to/outputs/out.biom", "/path/to/outputs/",
"/path/to/inputs/mapfile.txt", 'allgraphs')

# example with sampdepth argument for running allgraphs
trygraphwrapper("/path/to/outputs/out.biom", "/path/to/outputs/",
"/path/to/inputs/mapfile.txt", 'allgraphs', sampdepth=30000)


# example with optional argument sampdepth and tsv file
trygraphwrapper("/path/to/outputs/OTU_table.txt", "/path/to/outputs/",
"/path/to/inputs/mapfile.txt", 'allgraphs', sampdepth = 30000, tsvfile=TRUE)

# example of making heatmap with optional arguments
trygraphwrapper("/path/to/outputs/taxa_species.biom", "/path/to/outputs",
"/path/to/inputs/mapfile.txt", 'morphheatmap', sampdepth = 30000, filter_level=0.01,
taxlevel=c("Family", "seq"))
}


}
