% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGrid.R
\name{plotlyGrid}
\alias{plotlyGrid}
\title{Add Plotly data export to Plotly graph}
\source{
[plotlyGrid.R](../R/plotlyGrid.R)
}
\usage{
plotlyGrid(pplot, filename, data = NULL, title = NULL, outlib = "lib")
}
\arguments{
\item{pplot}{plotly or ggplot object}

\item{filename}{output filename (fullpath)}

\item{data}{data frame to export to plotly grid (optional for plotlyGrid)}

\item{title}{title of html page}

\item{outlib}{(Optional) name of external lib directory for non-selfcontained html.
Useful for multiple graphs sharing the same lib.}
}
\value{
html plot is saved to filename. external libraries are saved to outlib in same directory as filename.
Invisibly returns the plotly html widget.
}
\description{
All functions create an output html plot  with link which sends the data to a grid in the plotly chart
 studio.

 \code{plotlyGrid} takes in a ggplot or plotly object and creates an output html plotly plot.
}
